+++ mpc4j-s2pc-pjc/src/main/java/edu/alibaba/mpc4j/s2pc/pjc/main/pmid/PmidMain.java
@@ -200,7 +200,7 @@ public class PmidMain {
         LOGGER.info("(warmup) {} finish", pmidServer.ownParty().getPartyName());
     }
 
-    private void runServer(Rpc serverRpc, Party clientParty, PmidConfig config, long taskId, boolean parallel,
+    private void runServer(Rpc serverRpc, Party clientParty, PmidConfig config, int taskId, boolean parallel,
                            Map<ByteBuffer, Integer> serverElementMap, int serverU, int clientSetSize, int clientU,
                            PrintWriter printWriter) throws MpcAbortException {
         int serverSetSize = serverElementMap.keySet().size();
@@ -382,7 +382,7 @@ public class PmidMain {
         LOGGER.info("(warmup) {} finish", pmidClient.ownParty().getPartyName());
     }
 
-    private void runClient(Rpc clientRpc, Party serverParty, PmidConfig config, long taskId, boolean parallel,
+    private void runClient(Rpc clientRpc, Party serverParty, PmidConfig config, int taskId, boolean parallel,
                            Map<ByteBuffer, Integer> clientElementMap, int clientU, int serverSetSize, int serverU,
                            PrintWriter printWriter) throws MpcAbortException {
         int clientSetSize = clientElementMap.keySet().size();
