+++ mpc4j-s2pc-pso/src/test/java/edu/alibaba/mpc4j/s2pc/pso/oprp/OprpTest.java
@@ -158,7 +158,7 @@ public class OprpTest {
     }
 
     private void testPto(OprpSender sender, OprpReceiver receiver, int batchSize) {
-        long randomTaskId = Math.abs(SECURE_RANDOM.nextLong());
+        int randomTaskId = Math.abs(SECURE_RANDOM.nextInt());
         sender.setTaskId(randomTaskId);
         receiver.setTaskId(randomTaskId);
         try {
