+++ mpc4j-sml-opboost/src/test/java/edu/alibaba/mpc4j/sml/opboost/grad/OpGbdtRegMultiSlaveTest.java
@@ -191,7 +191,7 @@ public class OpGbdtRegMultiSlaveTest {
                                                DataFrame hostDataFrame, RegOpGradBoostHostConfig hostConfig,
                                                DataFrame leftSlaveDataFrame, OpBoostSlaveConfig leftSlaveConfig,
                                                DataFrame rightSlaveDataFrame, OpBoostSlaveConfig rightSlaveConfig) {
-        long randomTaskId = Math.abs(OpBoostTestUtils.SECURE_RANDOM.nextLong());
+        int randomTaskId = Math.abs(OpBoostTestUtils.SECURE_RANDOM.nextInt());
         host.setTaskId(randomTaskId);
         leftSlave.setTaskId(randomTaskId);
         rightSlave.setTaskId(randomTaskId);
