+++ mpc4j-sml-opboost/src/test/java/edu/alibaba/mpc4j/sml/opboost/grad/OpGbdtRegSingleSlaveTest.java
@@ -175,7 +175,7 @@ public class OpGbdtRegSingleSlaveTest {
     private GradientTreeBoost federateTraining(Formula formula,
                                                DataFrame hostDataFrame, RegOpGradBoostHostConfig hostConfig,
                                                DataFrame slaveDataFrame, OpBoostSlaveConfig slaveConfig) {
-        long randomTaskId = Math.abs(OpBoostTestUtils.SECURE_RANDOM.nextLong());
+        int randomTaskId = Math.abs(OpBoostTestUtils.SECURE_RANDOM.nextInt());
         host.setTaskId(randomTaskId);
         slave.setTaskId(randomTaskId);
         try {
