+++ mpc4j-sml-opboost/src/test/java/edu/alibaba/mpc4j/sml/opboost/xgboost/OpXgBoostRegMultiSlaveTest.java
@@ -262,7 +262,7 @@ public class OpXgBoostRegMultiSlaveTest {
                                        DataFrame leftSlaveDataFrame, OpBoostSlaveConfig leftSlaveConfig,
                                        DataFrame rightSlaveDataFrame, OpBoostSlaveConfig rightSlaveConfig
     ) {
-        long randomTaskId = Math.abs(OpBoostTestUtils.SECURE_RANDOM.nextLong());
+        int randomTaskId = Math.abs(OpBoostTestUtils.SECURE_RANDOM.nextInt());
         host.setTaskId(randomTaskId);
         leftSlave.setTaskId(randomTaskId);
         rightSlave.setTaskId(randomTaskId);
